backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
werkzeug
pytesseract
Pillow
pdf2image
python-dotenv
sqlalchemy
